Fuzzing: more help in finding ref-count problems
At the end of a run, try to free everything, to uncover ref-counting
problems.

If a problem is found, you can now run with $FUZZ_FREE set to a step
number. After that step, it will try to free everything. This can be
used to find which step corrupted the counts quickly by bisecting.

Also, split out some modules and added more state to the dump display.