Fix bug when handling a disembargo request to a switchable
If the client asked the server to echo back a disembargo request that
pointed at a client-hosted promise, the service would reject it,
claiming that the target was invalid. This is because a switchable
should not expose its current target, as it may change, but for
disembargoes this is OK. The solution is to move switchable to the
CapTP module and have it respond to the sealed CapTP message directly.