Fix cycle detector
Previously, attempting to resolve a promise to a set of fields that
contained a cycle would resolve the promise itself to a cycle error.

Now, we only break the individual fields that contain cycles.

One reason for doing this is because unnecessarily breaking the other
caps may affect some network paths and not others, leading to messages
being delivered along one path and not along another.

Also, add a `#problem` method to caps to test if they're broken.

Detected by the fuzz tests.