Autogenerate unit-tests for fuzzing failures
When examining a crash found by AFL, output a unit-test for it that can
be added to the test suite (with a little fixing up by hand).

This makes it easier to turn bugs found by the fuzzer into test-cases.

Added a test case that this generated (commented out because it is
actually failing at the moment).

Also, improved the logging output from the fuzzing in various ways.