Replace DynArray with WrapArray in most places
DynArray makes it hard to simplify example cases, because removing one
operation that added something to the array changes the positions in the
array of all future allocations, so all later operations do something
different.

Instead, let AFL pick the indexes to be used. If it picks one that's
already used, free the existing item first.