Forward pipelined calls on answers to the returned destination
Forward pipelined calls on answers to the returned destination, even if
we now know a shorter route. The spec requires this to ensure that
messages always arrive in order.

When we get a disembargo response, check what the new target is. If it
is a resolved question, act as if the reference now pointed at the
remote object and embargo again if necessary. Note: we will probably
need something similar to handle imports.