Clear blockers before forwarding calls on resolve
When a local promise resolves, we must notify anything waiting for us
that we're no longer a blocker and also forward the queued messages.

Before, we used a single queue for both. This could result in a promise
that was in the process of being resolved appearing to be settled, and
being reported as SenderHosted instead of SenderPromise.

Now, we have a separate list of promises we're blocking and we notify
them of the resolution first.

Found by AFL and simplified to new test case.