Separate steps of creating a client vat and connecting
Rename `Vat.connect` to `Vat.add_connection` and rename `Vat.live` to
`Vat.connect`. This seems clearer.

Replace `Capnp_rpc_unix.connect` with `Capnp_rpc_unix.client_only_vat`,
which just creates the client vat but doesn't connect it to anything.
The user must then use the new `Vat.connect` function to connect to
something.

This is more explicit and works better if the client needs to connect to
multiple services, or to connect multiple times to a single service (in
these cases, you only want to create a single vat).