Start with encryption on in the tutorial
The `Vat_config` API has changed slightly to make this easier.

The `--secret-key-type` option has gone. A vat must always have a key
now, although it may be ephemeral and is generated lazily. This is
because even if a vat accepts non-TLS incoming connections, it may still
need a key for out-bound connections.

Also, turn on `SO_REUSEADDR` for listening sockets so that the server
can always be restarted.