Secrets in sturdy refs
Sturdy refs now include a service ID component, which is passed in the
Cap'n Proto `Bootstrap` message. This field is marked as "deprecated"
in the spec, but there doesn't seem to be a replacement yet, it does
what we need, and the other libraries still use it.

Service IDs can be generated randomly or derived from another secret
(e.g. the vat's private key). The user can also supply a string
manually, e.g. for interoperability with systems with non-secret IDs.

Instead of passing an `~offer` argument with the public service, use
`~restore` to pass a `Restorer.t`.

Use `Vat.sturdy_ref` instead of `Vat.bootstrap_ref` or
`Vat.pp_bootstrap_uri`.

This system is intended to be extended later to support the persistence
system.

Also, add a TOC to README.