Add support for the persistence spec
`Persistence.with_persistence` can be used to extend any service
implementation with persistence support. `Persistence.with_sturdy_ref`
provides a simple wrapper for this for the common case where a service
knows its sturdy ref at the start.

`Persistence.save` or `save_exn` can be used to request a sturdy ref.

Added `Restorer.Table.sturdy_ref` to get a sturdy ref for a service in
the table. `Restorer.Table.create` now takes a function for generating
URIs from IDs so that in-memory-only tables can also support this.

Added `Sturdy_ref.cast`. This is needed when implementing a loader,
since the restorer doesn't know the type of the sturdy ref that it
passes in.