TLA+ model for restorer
Investigation of how the planned Lwt changes will affect existing code,
by making a model of one function.

Conclusion: the new semantics would break one or maybe two implicit
assumptions in this function and would result in a race condition,
allowing user code to receive an object with a ref-count of zero.