Reject attempts to send after disconnecting
When shutting down a connection we first turn every pending question
into an error (notifying the application) and then break all imported
caps.

If the application responds to the error by asking another question
before we break the caps, then we added the new question to the table.
Since we'd already cleared that table, we wouldn't do it again and the
question would never resolve.

Now, we refuse to add new questions if the connection is shutting down,
and instead return an error immediately.