Don't GC imports with callbacks
Normally, there are two references to an import: the application's
reference and the reference that the network system uses to update it
when it becomes resolved.

The reference held by the network system is a weak ref so that we can
detect if the application leaks the import without releasing it, and log
a warning about that.

However, it is possible that an application is only interested in an
import if it becomes resolved. In this case, the application's reference
to the import is only via a callback function attached to the import
itself. The result is that the import may be released early (with a
warning logged) and the application will find itself holding a broken
reference. This is likely to happen when using
`Capability.wait_until_settled`.