Prevent crash if ExtUnix wasn't build with sockopt
If ExtUnix isn't build with sockopt then the symbol `have_sockopt_int`
isn't available and an exception is raised.

This may be the case on some environments (e.g., ExtUnix build with
mingw-w64). It is a workaround until support for TCP_KEEPIDLE is added
for these environments (e.g., using Winsock instead).