Add connection progress indicator
Clients can now use `Capnp_rpc_unix.with_cap_exn` to show a progress
indicator while waiting to connect. It will use the log if enabled.
If not, but stderr is a tty, it shows a message while connecting
and then erases it (if connecting takes longer than 0.5s).
If stderr is not a tty, it just prints a message there.