Handle vat shutdown while connecting
When a vat is shut down it ends all registered connections and stops
accepting new ones. However, if a connection was in progress (e.g. doing
the TLS handshake but not yet registered) then we would miss it.

The `test_parallel_fails` test got upgraded to the stricter
`await_settled_exn` in #233, which was wrong but that wasn't detected
due to this bug.