Rename capnp-rpc-lwt package to capnp-rpc
The original idea was to allow different async libraries to be used, but
with effects we should only ever need one library, so it makes sense to
use the shorter `Capnp_rpc` name for this. The core protocol is now a
sub-library of the same package (`capnp-rpc.proto`). The protocol
modules that are generally useful are now re-exported by the new
`Capnp_rpc`.

This change is useful on its own, but the main aim is to reduce the diff
when switching from Lwt to Eio.